+++ lib/pedal_pico/pedal_pico_reverb/pedal_pico_reverb.c
@@ -42,8 +42,8 @@ void pedal_pico_reverb_process(int32_t normalized_1, uint16_t conversion_2, uint
     int32_t delay_1 = (int32_t)pedal_pico_reverb_delay_array[((pedal_pico_reverb_delay_index + PEDAL_PICO_REVERB_DELAY_TIME_MAX) - pedal_pico_reverb_delay_time_interpolation) % PEDAL_PICO_REVERB_DELAY_TIME_MAX];
     if (pedal_pico_reverb_delay_time_interpolation == 0) delay_1 = 0; // No Reverb, Otherwise Latest
     int32_t pedal_pico_reverb_normalized_1_amplitude = 0x00010000 - pedal_pico_reverb_delay_amplitude;
-    int32_t reduced_1 = (int32_t)(int64_t)((((int64_t)normalized_1 << 16) * (int64_t)pedal_pico_reverb_normalized_1_amplitude) >> 32);
-    delay_1 = (int32_t)(int64_t)((((int64_t)delay_1 << 16) * (int64_t)pedal_pico_reverb_delay_amplitude) >> 32);
+    int32_t reduced_1 = (int32_t)((((int64_t)normalized_1 << 16) * (int64_t)pedal_pico_reverb_normalized_1_amplitude) >> 32);
+    delay_1 = (int32_t)((((int64_t)delay_1 << 16) * (int64_t)pedal_pico_reverb_delay_amplitude) >> 32);
     int32_t mixed_1 = reduced_1 + delay_1;
     if (sw_mode == 1) {
         pedal_pico_reverb_delay_array[pedal_pico_reverb_delay_index] = (int16_t)normalized_1;
