+++ lib/pedal_pico/pedal_pico_tape/pedal_pico_tape.c
@@ -57,8 +57,8 @@ void pedal_pico_tape_process(int32_t normalized_1, uint16_t conversion_2, uint16
     int32_t delay_1 = (int32_t)pedal_pico_tape_delay_array[((pedal_pico_tape_delay_index + PEDAL_PICO_TAPE_DELAY_TIME_MAX) - ((int16_t)pedal_pico_tape_delay_time + time_swing)) % PEDAL_PICO_TAPE_DELAY_TIME_MAX];
     if (pedal_pico_tape_delay_time + time_swing == 0) delay_1 = 0; // No Delay, Otherwise Latest
     int32_t pedal_pico_tape_normalized_1_amplitude = 0x00010000 - pedal_pico_tape_delay_amplitude;
-    normalized_1 = (int32_t)(int64_t)((((int64_t)normalized_1 << 16) * (int64_t)pedal_pico_tape_normalized_1_amplitude) >> 32);
-    delay_1 = (int32_t)(int64_t)((((int64_t)delay_1 << 16) * (int64_t)pedal_pico_tape_delay_amplitude) >> 32);
+    normalized_1 = (int32_t)((((int64_t)normalized_1 << 16) * (int64_t)pedal_pico_tape_normalized_1_amplitude) >> 32);
+    delay_1 = (int32_t)((((int64_t)delay_1 << 16) * (int64_t)pedal_pico_tape_delay_amplitude) >> 32);
     int32_t mixed_1 = normalized_1 + delay_1;
     pedal_pico_tape_delay_array[pedal_pico_tape_delay_index] = (int16_t)mixed_1;
     pedal_pico_tape_delay_index++;
